separate pango and fontconfig dependencies
